+++ src/main/java/com/gabrielhenrique/salesapicomeia/modules/sales/repository/SalesRepository.java
@@ -1,9 +1,11 @@
-package com.gabrielhenrique.salesapicomeia.modules.sales;
+package com.gabrielhenrique.salesapicomeia.modules.sales.repository;
 
 import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.gabrielhenrique.salesapicomeia.modules.sales.entity.SalesEntity;
 
 public interface SalesRepository extends JpaRepository<SalesEntity, UUID> {
     Optional<SalesEntity> findBySaleDescription(String saleDescription);
